backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.4
alembic==1.13.1

# AI/ML
openai==1.12.0
google-generativeai>=0.8.0
tiktoken==0.6.0

# Document processing
pypdf2==3.0.1
pdfplumber==0.10.3
markdown==3.5.2
beautifulsoup4==4.12.3
httpx==0.26.0

# Audio processing
pydub==0.25.1

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1
uuid6==2024.1.12

# For chunking
langchain-text-splitters==0.0.1
